+++ contracts/Shared.sol
@@ -8,7 +8,7 @@ library Shared {
         bool hidden;
         uint24 maxSupply; // can be minted up to MAX_SUPPLY
         uint24 royaltyFriction; // used for `royaltyInfo` (ERC2981) and `seller_fee_basis_points` (Opeansea's Contract-level metadata)
-        uint40 mintingBeginsFrom; // Timestamp that minting event begins
+        uint32 mintingBeginsFrom; // Timestamp that minting event begins
         uint256 mintingCost; // Native token (ETH, BNB, KLAY, etc)
         string description;
     }
